+++ src/media/FFmpegUtils.h
@@ -3,9 +3,9 @@
 #include "media.h"
 
 #define QmBaseTimeToSecondTime(value, base) (value * double(base.num) )/(base.den)
-#define QmSecondTimeToBaseTime(value, base) (value * double(base.den) )/(base.num)
+#define QmSecondTimeToBaseTime(value, base) (int64_t)((value * double(base.den) )/(base.num))
 #define QmBaseTimeToMSTime(value, base) int(1000 * (value * double(base.num) )/(base.den))
-#define QmMSTimeToBaseTime(value, base) (value * double(base.den) )/(base.num * 1000)
+#define QmMSTimeToBaseTime(value, base) (int64_t)((value * double(base.den) )/(base.num * 1000))
 
 struct AVRational;
 
@@ -17,6 +17,6 @@ public:
     static int fourccToFFmpegFormat(int);
     static int ffmpegFormatToFourcc(int);
 	static AVPacketPtr allocAVPacket();
-	static int64_t currentMilliSecsSinceEpoch();
+	static int currentMilliSecsSinceEpoch();
 };
